+++ src/main/kotlin/site/siredvin/turtlematic/api/AutomataCoreTier.kt
@@ -11,7 +11,7 @@ enum class AutomataCoreTier(
     TIER1(2, 2, 1.0),
     TIER2(4, 3, 1.0),
     TIER3(8, 4, 0.8, setOf(AutomataCoreTraits.DURABILITY_REFUND_CHANCE)),
-    TIER4(16, 6, 0.5, setOf(AutomataCoreTraits.DURABILITY_REFUND)),
+    TIER4(16, 6, 0.5, setOf(AutomataCoreTraits.DURABILITY_REFUND, AutomataCoreTraits.STARBOUND_REGENERATION)),
     ENORMOUS_TIER(Integer.MAX_VALUE, Integer.MAX_VALUE, 0.0, setOf(AutomataCoreTraits.DURABILITY_REFUND));
 
     private var _interactionRadius: ForgeConfigSpec.IntValue? = null