+++ src/main/kotlin/site/siredvin/turtlematic/api/AutomataCoreTraits.kt
@@ -6,4 +6,5 @@ import site.siredvin.turtlematic.Turtlematic
 object AutomataCoreTraits {
     val DURABILITY_REFUND_CHANCE = ResourceLocation(Turtlematic.MOD_ID, "durability_refund_chance")
     val DURABILITY_REFUND = ResourceLocation(Turtlematic.MOD_ID, "durability_refund")
+    val STARBOUND_REGENERATION = ResourceLocation(Turtlematic.MOD_ID, "starbound_regeneration")
 }