+++ src/main/kotlin/site/siredvin/turtlematic/util/TooltipHandlerCollection.kt
@@ -1,7 +1,6 @@
 package site.siredvin.turtlematic.util
 
 import net.minecraft.network.chat.Component
-import net.minecraft.network.chat.TextComponent
 import net.minecraft.world.item.Item
 import site.siredvin.lib.util.text
 import site.siredvin.turtlematic.Turtlematic
@@ -24,21 +23,23 @@ object TooltipHandlerCollection {
         return TOOLTIP_PROVIDERS_PER_ITEM[item]
     }
 
-    fun durabilityRefundChanceTooltip(coreTier: IAutomataCoreTier, tooltipList: MutableList<Component>) {
+    fun commonNetheriteTooltip(coreTier: IAutomataCoreTier, tooltipList: MutableList<Component>) {
         if (coreTier.traits.contains(AutomataCoreTraits.DURABILITY_REFUND_CHANCE))
             tooltipList.add(text(Turtlematic.MOD_ID, "durability_refund_chance"))
     }
 
-    fun durabilityRefundTooltip(coreTier: IAutomataCoreTier, tooltipList: MutableList<Component>) {
+    fun commonStarboundTooltip(coreTier: IAutomataCoreTier, tooltipList: MutableList<Component>) {
         if (coreTier.traits.contains(AutomataCoreTraits.DURABILITY_REFUND))
             tooltipList.add(text(Turtlematic.MOD_ID, "durability_refund"))
+        if (coreTier.traits.contains(AutomataCoreTraits.STARBOUND_REGENERATION))
+            tooltipList.add(text(Turtlematic.MOD_ID, "starbound_generation"))
     }
 
     fun registerDefaults() {
-        registerProvider(Items.NETHERITE_END_AUTOMATA_CORE, this::durabilityRefundChanceTooltip)
-        registerProvider(Items.NETHERITE_HUSBANDRY_AUTOMATA_CORE, this::durabilityRefundChanceTooltip)
-        registerProvider(Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE, this::durabilityRefundTooltip)
-        registerProvider(Items.STARBOUND_END_AUTOMATA_CORE, this::durabilityRefundTooltip)
-        registerProvider(Items.ENORMOUS_AUTOMATA_CORE, this::durabilityRefundTooltip)
+        registerProvider(Items.NETHERITE_END_AUTOMATA_CORE, this::commonNetheriteTooltip)
+        registerProvider(Items.NETHERITE_HUSBANDRY_AUTOMATA_CORE, this::commonNetheriteTooltip)
+        registerProvider(Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE, this::commonStarboundTooltip)
+        registerProvider(Items.STARBOUND_END_AUTOMATA_CORE, this::commonStarboundTooltip)
+        registerProvider(Items.ENORMOUS_AUTOMATA_CORE, this::commonStarboundTooltip)
     }
 }