+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to harvest crops for husbandry automata
 - Extra information about crops for husbandry automata
 - Piston and sticky piston turtles
+- Traits description on automata cores
+- Base automata levels: netherite, starbound and creative
 
 ### Changed